Formatted and moved paths.
Paths are now based on &str.
