+++ src/dir.rs
@@ -1,5 +1,21 @@
 //! Directory related types
 
+// ============================================================================
+// Imports
+// ============================================================================
+
+// None
+
+// ============================================================================
+// Constants
+// ============================================================================
+
+// None
+
+// ============================================================================
+// Types
+// ============================================================================
+
 /// Represents an open directory
 #[repr(C)]
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
@@ -8,12 +24,9 @@ pub struct Handle(u8);
 impl Handle {
     /// Construct a new `Handle` from an integer.
     ///
-    /// Only the OS should call this.
-    ///
-    /// # Safety
-    ///
-    /// The integer given must be a valid index for an open Directory.
-    #[cfg(feature = "os")]
+    /// Only the OS should call this - applications should not be constructing
+    /// their own dir handles! But if you do, you probably can't harm anything
+    /// and it's no worse that C just using `int`.
     pub const fn new(value: u8) -> Handle {
         Handle(value)
     }
@@ -30,8 +43,28 @@ impl Handle {
 #[repr(C)]
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Entry {
-    /// The name and extension of the file
+    /// The name and extension of the file.
+    ///
+    /// The name and extension are separated by a single '.'.
+    ///
+    /// The filename will be in ASCII. Unicode filenames are not supported.
     pub name: [u8; crate::MAX_FILENAME_LEN],
-    /// File properties
+    /// The properties for the file/directory this entry represents.
     pub properties: crate::file::Stat,
 }
+
+// ============================================================================
+// Functions
+// ============================================================================
+
+// None
+
+// ============================================================================
+// Tests
+// ============================================================================
+
+// None
+
+// ============================================================================
+// End of File
+// ============================================================================
